Кнопка сброса работает на фильтры карты, доработка сброса/отправки формы по тех.заданию
